added readme and consolidated files in infra repo
