+++ apps/igloo-kit-cli/src/framework.rs
@@ -1,28 +1,7 @@
-use std::path::PathBuf;
-
 use clap::Subcommand;
 
 pub mod directories;
 pub mod schema;
-
-enum Languages {
-    Python,
-    Typescript,
-}
-
-struct Template {
-    name: String,
-    language: Languages,
-    path: PathBuf, // The path to the local file
-    // remote_path: Url, // The path to the remote file
-}
-
-trait FrameworkObject {
-    fn new() -> Self;
-    fn directory(&self) -> PathBuf;
-    fn templates(&self) -> Vec<Template>;
-}
-
 pub enum Insights {
     Metric,
     Dashboard,
