--- apps/igloo-kit-cli/src/framework/schema/templates/clickhouse/mod.rs
@@ -1,1 +1,0 @@
-pub mod sql;
