--- apps/igloo-kit-cli/src/framework/schema/templates/mod.rs
@@ -1,2 +1,0 @@
-pub mod sql;
-pub mod clickhouse;
