+++ apps/igloo-kit-cli/src/infrastructure/olap/clickhouse.rs
@@ -1,22 +1,16 @@
+pub mod config;
+pub mod mapper;
+mod queries;
+
 use std::fmt;
 
 use clickhouse::Client;
 use reqwest::Url;
 use schema_ast::ast::FieldArity;
 
-use crate::framework::schema::{OpsTable, templates::clickhouse::sql::{CreateTableQuery, DropTableQuery}, UnsupportedDataTypeError};
+use crate::framework::schema::{OpsTable, UnsupportedDataTypeError};
 
-#[derive(Debug, Clone)]
-pub struct ClickhouseConfig {
-    pub db_name: String, // ex. local
-    pub user: String,
-    pub password: String,
-    pub host: String, // ex. localhost
-    pub host_port: i32, // ex. 18123
-    pub postgres_port: i32, // ex. 9005
-    pub kafka_port: i32, // ex. 9092
-    pub cluster_network: String, // ex. panda-house
-}
+use self::{queries::{CreateTableQuery, DropTableQuery}, config::ClickhouseConfig};
 
 #[derive(Debug, Clone)]
 pub enum ClickhouseTableType {
