+++ apps/igloo-kit-cli/src/infrastructure/olap/clickhouse/config.rs
@@ -1,40 +1,62 @@
-use std::fs::File;
-use std::io::prelude::*;
-use std::path::PathBuf;
+//! # Clickhouse Config
+//! Module to handle the creation of the Clickhouse config files
+//! 
+//! ## Suggested Improvements
+//! - we need to understand clickhouse configuration better before we can go deep on it's configuration
+//! 
 
-pub fn create_server_config_file(directory: &PathBuf) -> std::io::Result<()> {
-    let mut file = File::create(directory.join("psql.xml"))?;
-    file.write_all(b"<clickhouse><postgresql_port>9005</postgresql_port></clickhouse>")?;
-
-    Ok(())
+#[derive(Debug, Clone)]
+pub struct ClickhouseConfig {
+    pub db_name: String, // ex. local
+    pub user: String,
+    pub password: String,
+    pub host: String, // ex. localhost
+    pub host_port: i32, // ex. 18123
+    pub postgres_port: i32, // ex. 9005
+    pub kafka_port: i32, // ex. 9092
+    pub cluster_network: String, // ex. panda-house
 }
 
-pub fn create_user_config_file(directory: &PathBuf) -> std::io::Result<()> {
-    let mut file = File::create(directory.join("panda.xml"))?;
-    file.write_all(b"
-    <clickhouse>
-        <profiles>
-            <experimental>
-                <allow_experimental_object_type>0</allow_experimental_object_type>
-            </experimental>
-        </profiles>
-        <users>
-            <panda>
-                <profile>experimental</profile>
-            </panda>
-        </users>
-    </clickhouse>
-    ")?;
+// *** The following configurations work. They just didn't solve the problem I needed them to solve. ***
+// 
+// We should continue to explore them 
+// 
+// use std::fs::File;
+// use std::io::prelude::*;
+// use std::path::PathBuf;
 
-    Ok(())
-}
+// pub fn create_server_config_file(directory: &PathBuf) -> std::io::Result<()> {
+//     let mut file = File::create(directory.join("psql.xml"))?;
+//     file.write_all(b"<clickhouse><postgresql_port>9005</postgresql_port></clickhouse>")?;
 
+//     Ok(())
+// }
 
-pub fn create_init_script(directory: &PathBuf) -> std::io::Result<()> {
-    let mut file = File::create(directory.join("db.sql"))?;
-    file.write_all(b"
-    SET allow_experimental_object_type = 1;
-    ")?;
+// pub fn create_user_config_file(directory: &PathBuf) -> std::io::Result<()> {
+//     let mut file = File::create(directory.join("panda.xml"))?;
+//     file.write_all(b"
+//     <clickhouse>
+//         <profiles>
+//             <experimental>
+//                 <allow_experimental_object_type>0</allow_experimental_object_type>
+//             </experimental>
+//         </profiles>
+//         <users>
+//             <panda>
+//                 <profile>experimental</profile>
+//             </panda>
+//         </users>
+//     </clickhouse>
+//     ")?;
 
-    Ok(())
-}
+//     Ok(())
+// }
+
+// pub fn create_init_script(directory: &PathBuf) -> std::io::Result<()> {
+//     let mut file = File::create(directory.join("db.sql"))?;
+//     file.write_all(b"
+//     SET allow_experimental_object_type = 1;
+//     ")?;
+
+//     Ok(())
+// }
